+++ studio-platform-user/src/main/java/studio/echo/base/user/web/dto/GroupDto.java
@@ -71,6 +71,6 @@ public class GroupDto {
 
     int roleCount = 0;
 
-    Integer memberCount ; 
+    Long memberCount ; 
 
 }
